fix(addressbook): crash

+++ src/atomic.dex.qt.addressbook.model.cpp
@@ -86,7 +86,7 @@ namespace atomic_dex
         for (int row = 0; row < rows; ++row)
         {
             contact_model* element = this->m_addressbook.at(position);
-            if (element->rowCount(QModelIndex()) == 0 || (this->m_should_delete_contacts && not element->get_name().isEmpty()))
+            if ((element->rowCount(QModelIndex()) == 0 && not element->get_name().isEmpty()) || (this->m_should_delete_contacts && not element->get_name().isEmpty()))
             {
                 this->m_wallet_manager.delete_contact(element->get_name());
                 this->m_wallet_manager.update_wallet_cfg();
